move landuse color table to appdata; shared by front and backend. #41
