+++ appdata/NLCD_2016.lulcdata.json
@@ -1,0 +1,18 @@
+{
+  "11": { "name": "Open Water", "color": "#466b9f" },
+  "12": { "name": "Perennial Ice/Snow", "color": "#d1def8" },
+  "21": { "name": "Developed, Open Space", "color": "#EBC7E6" },
+  "22": { "name": "Developed, Low Intensity", "color": "#BFACE2" },
+  "23": { "name": "Developed, Medium Intensity", "color": "#A084DC" },
+  "24": { "name": "Developed, High Intensity", "color": "#645CBB" },
+  "31": { "name": "Barren Land (Rock/Sand/Clay)", "color": "#b3ac9f" },
+  "41": { "name": "Deciduous Forest", "color": "#68ab5f" },
+  "42": { "name": "Evergreen Forest", "color": "#1c5f2c" },
+  "43": { "name": "Mixed Forest", "color": "#b5c58f" },
+  "52": { "name": "Shrub/Scrub", "color": "#ccb879" },
+  "71": { "name": "Grassland/Herbaceous", "color": "#dfdfc2" },
+  "81": { "name": "Pasture/Hay", "color": "#dcd939" },
+  "82": { "name": "Cultivated Crops", "color": "#ab6c28" },
+  "90": { "name": "Woody Wetlands", "color": "#b8d9eb" },
+  "95": { "name": "Emergent Herbaceous Wetlands", "color": "#6c9fb8" }
+}
