+++ frontend/src/index.css
@@ -183,9 +183,11 @@ html, body, #root, .App {
   width: 200px;
 }
 
+/*de-saturate to match contrast adjustment in lulcLayer.js*/
 .study-area-table .parcel-block {
   display: flex;
   flex-wrap:  wrap;
+  filter: saturate(40%);
 }
 
 .study-area-table .hidden-row {
@@ -211,6 +213,7 @@ html, body, #root, .App {
   padding: 0.5rem;
   border-left-style: solid;
   border-left-width: 8px;
+  filter: saturate(40%);
 }
 
 .scenario-table .table-note {
