--- frontend/src/landuseCodes.js
@@ -1,20 +1,0 @@
-// https://www.mrlc.gov/data/legends/national-land-cover-database-class-legend-and-description
-// Here we exclude 4 Alaska-only types
-export default {
-  11: { name: 'Open Water', color: '#466b9f' },
-  12: { name: 'Perennial Ice/Snow', color: '#d1def8' },
-  21: { name: 'Developed, Open Space', color: '#dec5c5' },
-  22: { name: 'Developed, Low Intensity', color: '#d99282' },
-  23: { name: 'Developed, Medium Intensity', color: '#eb0000' },
-  24: { name: 'Developed, High Intensity', color: '#ab0000' },
-  31: { name: 'Barren Land (Rock/Sand/Clay)', color: '#b3ac9f' },
-  41: { name: 'Deciduous Forest', color: '#68ab5f' },
-  42: { name: 'Evergreen Forest', color: '#1c5f2c' },
-  43: { name: 'Mixed Forest', color: '#b5c58f' },
-  52: { name: 'Shrub/Scrub', color: '#ccb879' },
-  71: { name: 'Grassland/Herbaceous', color: '#dfdfc2' },
-  81: { name: 'Pasture/Hay', color: '#dcd939' },
-  82: { name: 'Cultivated Crops', color: '#ab6c28' },
-  90: { name: 'Woody Wetlands', color: '#b8d9eb' },
-  95: { name: 'Emergent Herbaceous Wetlands', color: '#6c9fb8' },
-};
